+++ sync-for-expenses/RELEASES.md
@@ -393,3 +393,13 @@ Based on:
 - [go v4.0.0] sync-for-expenses
 ### Releases
 - [Go v4.0.0] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v4.0.0 - sync-for-expenses
+
+## 2023-12-04 01:25:32
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v4.0.1] sync-for-expenses
+### Releases
+- [Go v4.0.1] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v4.0.1 - sync-for-expenses